requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
curl_cffi==0.14.0
fastapi==0.115.6
Flask==3.1.2
Flask-SocketIO==5.5.1
frozendict==2.4.7
greenlet==3.3.0
h11==0.16.0
idna==3.11
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
line_profiler==5.0.0
mando==0.7.1
MarkupSafe==3.0.3
more-itertools==10.8.0
multitasking==0.0.12
nodeenv==1.9.1
nodejs-wheel-binaries==24.12.0
numpy==2.3.5
pandas==2.3.3
peewee==3.18.3
platformdirs==4.5.1
protobuf==6.33.2
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyright==1.1.407
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.15.1
pytz==2025.2
PyYAML==6.0.3
radon==6.0.1
requests==2.32.5
ruff==0.14.10
simple-websocket==1.1.0
six==1.17.0
snakeviz==2.2.2
soupsieve==2.8.1
sqlacodegen==3.2.0
SQLAlchemy==2.0.45
toml==0.10.2
tornado==6.5.4
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.34.0
websockets==15.0.1
Werkzeug==3.1.4
wsproto==1.3.2
yfinance==0.2.66
